Move hilbert transformation back to quatrex
- The hilbert transformation a specific for polarization and the GW self-energy. Thus, the should be located in the same file.
- Move the tests accordingly.
- Move the `_naive_convolution` from the tests to qttools as it is now used in multiple tests.
