+++ src/qttools/convolutions/ffts.py
@@ -125,89 +125,3 @@ def fft_correlate_kpoints(a: NDArray, b: NDArray) -> NDArray:
         shift=1,
         axis=tuple(range(1, len(nka) + 1)),
     )
-
-
-@profiler.profile(level="api")
-def hilbert_transform_polarization(a: NDArray, energies: NDArray) -> NDArray:
-    """Computes the Hilbert transform of the array a, assuming the symmetries of the
-    polarization, i.e \([P^{\lessgtr}_{ij}(\omega)]^{\dagger} = -P^{\gtrless}_{ij}(-\omega)\).
-    This becomes \(a(-\omega)=a^{*}(\omega)\), where a is \(a=P^>-P^<\).
-
-    Assumes that the first axis corresponds to the energy axis.
-
-    Parameters
-    ----------
-    a : NDArray
-        The array to transform.
-    energies : NDArray
-        The energy values corresponding to the first axis of a.
-    eta : float, optional
-        For the principle part. Small part to avoid singularity, by
-        default 1e-8.
-
-    Returns
-    -------
-    NDArray
-         The Hilbert transform of a.
-
-    """
-    # eta for removing the singularity. See Cauchy principal value.
-    de = energies[1] - energies[0]
-    eta = de / 2
-    energy_differences = (
-        xp.expand_dims(energies - energies[0], tuple(range(1, a.ndim))) + eta
-    )
-    ne = energies.size
-
-    hilbert_kernel = 1 / energy_differences
-    b = fft_convolve(a, hilbert_kernel)[:ne]
-    # Negative frequencies of a
-    b += fft_convolve(a[::-1].conj(), hilbert_kernel)[-ne:]
-    # Negative frequencies of the kernel
-    hilbert_kernel = -hilbert_kernel[::-1]
-    b += fft_convolve(a, hilbert_kernel)[-ne:]
-
-    return b
-
-
-@profiler.profile(level="api")
-def hilbert_transform_selfenergy(
-    sl: NDArray, sg: NDArray, energies: NDArray
-) -> NDArray:
-    """Computes the Hilbert transform.
-
-    Assumes that the first axis corresponds to the energy axis.
-
-    Parameters
-    ----------
-    sl : NDArray
-        The lesser self-energy on the grid |-----|-----|xxxxx|.
-    sg : NDArray
-        The greater self-energy on the grid |xxxxx|-----|-----|.
-    energies : NDArray
-        The energy values corresponding to the first axis of a.
-
-    Returns
-    -------
-    NDArray
-         The Hilbert transform of a.
-
-    """
-    ne = energies.size
-    nk = sg.shape[1:-1]
-    # Add empty dimensions for each k-point.
-    energy_differences = (energies - energies[0]).reshape(-1, *(len(nk) + 1) * (1,))
-    # eta for removing the singularity. See Cauchy principal value.
-    eta = (energies[1] - energies[0]) / 2
-    hilbert_kernel = 1 / (energy_differences + eta)
-
-    sr = fft_convolve(sg[:ne] - sl[-ne:], hilbert_kernel)[:ne]
-    # Correct for left edge
-    sr += fft_convolve(-sl[:ne], hilbert_kernel)[-ne:]
-    # Next account for negative frequencies
-    hilbert_kernel = -hilbert_kernel[::-1]
-    sr += fft_convolve(sg[:ne] - sl[-ne:], hilbert_kernel)[-ne:]
-    # Correct for right edge
-    sr += fft_convolve(sg[-ne:], hilbert_kernel)[:ne]
-
-    return sr
